+++ tests/core/pyspec/eth2spec/test/bellatrix/block_processing/test_process_execution_payload.py
@@ -1,8 +1,8 @@
 from random import Random
 
-from eth2spec.debug.random_value import get_random_bytes_list
 from eth2spec.test.helpers.execution_payload import (
     build_empty_execution_payload,
+    build_randomized_execution_payload,
     get_execution_payload_header,
     build_state_with_incomplete_transition,
     build_state_with_complete_transition,
@@ -305,33 +305,6 @@ def test_zero_length_transaction_first_payload(spec, state):
 def test_zero_length_transaction_regular_payload(spec, state):
     state = build_state_with_complete_transition(spec, state)
     yield from run_zero_length_transaction_test(spec, state)
-
-
-def build_randomized_execution_payload(spec, state, rng):
-    execution_payload = build_empty_execution_payload(spec, state)
-    execution_payload.fee_recipient = spec.ExecutionAddress(get_random_bytes_list(rng, 20))
-    execution_payload.state_root = spec.Bytes32(get_random_bytes_list(rng, 32))
-    execution_payload.receipts_root = spec.Bytes32(get_random_bytes_list(rng, 32))
-    execution_payload.logs_bloom = spec.ByteVector[spec.BYTES_PER_LOGS_BLOOM](
-        get_random_bytes_list(rng, spec.BYTES_PER_LOGS_BLOOM)
-    )
-    execution_payload.block_number = rng.randint(0, 10e10)
-    execution_payload.gas_limit = rng.randint(0, 10e10)
-    execution_payload.gas_used = rng.randint(0, 10e10)
-    extra_data_length = rng.randint(0, spec.MAX_EXTRA_DATA_BYTES)
-    execution_payload.extra_data = spec.ByteList[spec.MAX_EXTRA_DATA_BYTES](
-        get_random_bytes_list(rng, extra_data_length)
-    )
-    execution_payload.base_fee_per_gas = rng.randint(0, 2**256 - 1)
-    execution_payload.block_hash = spec.Hash32(get_random_bytes_list(rng, 32))
-
-    num_transactions = rng.randint(0, 100)
-    execution_payload.transactions = [
-        spec.Transaction(get_random_bytes_list(rng, rng.randint(0, 1000)))
-        for _ in range(num_transactions)
-    ]
-
-    return execution_payload
 
 
 def run_randomized_non_validated_execution_fields_test(spec, state, execution_valid=True, rng=Random(5555)):
